Decided on a *much* better approach with parsing ELF and dwarf.
Since we *know* where the data is and we are in _full_ control of it
getting bogged down by Rust's borrowing system is unnecessary. Therefore
we wrap HashMap<SectionId, &[u8]> in a struct that instead has a map of
HashMap<SectionId, (*const u8, usize)> that, on the API boundary hands out slices of &[u8]. We can do this safely, because we know the data that it points to, never moves, and we are in full ctrl of it.