added "list lines" functionality to REPL client.
Currently, doesn't "exactly" work, because of this; finding the root compilation folder
is not currently implemented, so it just appends a source file to "./" when it sees it.

This attribute can be read from the compilation_unit's attribute tag DW_AT_name
thus, we should parse this information and hand it to the LineNumberProgram
before executing the line number program.